consolidate legacy static facade
