+++ misda/__init__.py
@@ -9,9 +9,6 @@ MISDA is a graph-theoretic framework designed for dimensionality reduction in Mu
 
 import numpy as np
 import pandas as pd
-from scipy import stats
-import networkx as nx
-import matplotlib.pyplot as plt
 import math
 from dataclasses import dataclass, field
 from typing import Optional, Tuple, Dict, List, Any
@@ -27,6 +24,7 @@ from ._statistics import (
     AlphaRegime,
     diagnose_alpha_regime,
     describe_alpha_regime,
+    calculate_spectral_entropy,
 )
 from ._graph import (
     find_maximal_independent_sets,
@@ -49,6 +47,14 @@ from ._pareto import (
     evaluate_pareto_raw,
 )
 from ._validation import _validate_input_matrix
+from ._metadata import __version__
+from ._plotting import (
+    _enforce_min_distance,
+    _parse_node_to_1based,
+    _extract_mis_nodes_1based,
+    plot_custom_misda_graph,
+)
+from ._reporting import explain_ses
 from .result import MISCandidate, MISDAResult
 from .api import (
     report_significant_correlations,
@@ -57,8 +63,7 @@ from .api import (
     _analyze_static_fast,
     _analyze_static,
 )
-
-__version__ = "0.4.2"
+from .benchmark import compile_benchmark_summary
 
 # Constants
 AGGRESSIVE = 0
@@ -68,340 +73,15 @@ CONSERVATIVE = 1
 # Utilities
 
 
-def _enforce_min_distance(pos, min_dist=0.28, iters=900, jitter=1e-3, seed=7):
-    """Adjusts 2D layout positions to enforce a minimum distance between nodes."""
-    rng = np.random.default_rng(seed)
-    nodes = list(pos.keys())
-    if not nodes:
-        return pos
-
-    P = np.array([pos[n] for n in nodes], dtype=float)
-    P += 1e-12 * rng.normal(size=P.shape)
-
-    for _ in range(iters):
-        moved = False
-        for i in range(len(nodes)):
-            for j in range(i + 1, len(nodes)):
-                d = P[j] - P[i]
-                dist = float(np.hypot(d[0], d[1]))
-                if dist < 1e-12:
-                    P[j] += rng.normal(scale=jitter, size=2)
-                    moved = True
-                elif dist < min_dist:
-                    push = d / dist
-                    delta = 0.5 * (min_dist - dist) * push
-                    P[i] -= delta
-                    P[j] += delta
-                    moved = True
-        if not moved:
-            break
-    return {n: P[k] for k, n in enumerate(nodes)}
-
-
-def _parse_node_to_1based(x, M):
-    """Accepts 0-based int, 1-based int, 'fK', and 'K'."""
-    if isinstance(x, (int, np.integer)):
-        xi = int(x)
-        if 0 <= xi < M:
-            return xi + 1
-        if 1 <= xi <= M:
-            return xi
-        return None
-
-    s = str(x).strip()
-    if len(s) >= 2 and s[0] in ("f", "F"):
-        s = s[1:]
-
-    try:
-        xi = int(s)
-    except Exception:
-        return None
-
-    if 0 <= xi < M:
-        return xi + 1
-    if 1 <= xi <= M:
-        return xi
-    return None
 
 
 
-def calculate_spectral_entropy(Y):
-    """
-    Calculates the normalized spectral entropy of the correlation matrix of Y.
-    High entropy (~1.0) indicates complex, spherical, or random structure.
-    Low entropy (~0.0) indicates high redundancy/dimensionality reduction potential.
-    """
-    if hasattr(Y, "values"):
-        data = np.asarray(Y.values, dtype=float)
-    else:
-        data = np.asarray(Y, dtype=float)
-    
-    n, m = data.shape
-    if m < 2:
-        return 0.0
-        
-    # Correlation matrix
-    corr = np.corrcoef(data, rowvar=False)
-    # Eigenvalues (Hermitian/Symmetric)
-    eigvals = np.linalg.eigvalsh(corr)
-    
-    # Normalize eigenvalues to probability distribution
-    # Filter small negative/zeros due to precision
-    eigvals = eigvals[eigvals > 1e-9]
-    if len(eigvals) == 0:
-        return 0.0
-        
-    p = eigvals / np.sum(eigvals)
-    
-    # Entropy
-    se = -np.sum(p * np.log(p))
-    
-    # Normalize by log(M)
-    # Note: Max entropy for M variables is log(M) when all eigenvalues = 1
-    # However, number of non-zero eigenvalues could be < M if N < M.
-    # Usually we norm by log(min(N, M)) or log(len(eigvals)).
-    # Using log(len(eigvals)) is safer.
-    denom = np.log(len(eigvals))
-    if denom == 0:
-        return 0.0
-        
-    return se / denom
 
 
-def _extract_mis_nodes_1based(mis_entry, M):
-    """
-    Strict extractor (no random hunting):
-      - mis_indices: list of ints (0-based or 1-based)
-      - mis: list (ints/labels)
-      - mis_nodes: list (ints/labels)
-    Returns 1..M nodes (deduplicated, preserving order).
-    """
-    if not isinstance(mis_entry, dict):
-        raise ValueError(f"mis_ranked item is not a dict: {type(mis_entry)}")
-
-    raw = None
-    for k in ("mis_indices", "mis", "mis_nodes"):
-        if k in mis_entry and mis_entry[k] not in (None, [], ()):
-            raw = mis_entry[k]
-            break
-
-    if raw is None:
-        keys = sorted(mis_entry.keys())
-        raise ValueError(
-            "mis_ranked item does not contain MIS in any canonical key "
-            "('mis_indices', 'mis', 'mis_nodes'). "
-            f"Item keys: {keys}"
-        )
-
-    xs = list(raw) if isinstance(raw, (list, tuple, set)) else [raw]
-
-    out, seen = [], set()
-    for x in xs:
-        u = _parse_node_to_1based(x, M)
-        if u is None:
-            continue
-        if u not in seen:
-            seen.add(u)
-            out.append(u)
-
-    return out
 
 
-def plot_custom_misda_graph(
-    results: dict,
-    figsize=(10, 8),
-    min_dist=0.5,
-    title="ISDA Graph",
-    show_removed=True
-):
-    """
-    Plots the dependency graph derived from MISDA analysis.
-    Nodes are objectives, edges are significant correlations.
-    """
-    M = results["M"]
-    A = np.asarray(results.get("adjacency", None))
-    if A is None:
-        raise ValueError("results['adjacency'] missing.")
-
-    # Use actual labels if provided in the results dict
-    labels = results.get("labels")
-    if labels is not None:
-        nodes = list(labels)
-        # Map node name to 1-based index (internal ISDA logic uses 1-based)
-        node_to_idx = {name: i for i, name in enumerate(nodes)}
-    else:
-        nodes = list(range(1, M + 1))
-        node_to_idx = {i: i-1 for i in nodes}
-
-    # --- MIS: UNIQUE and explicit source (mis_ranked) ---
-    mis_ranked = results.get("mis_ranked", None)
-    if not isinstance(mis_ranked, list) or len(mis_ranked) == 0:
-        raise ValueError(
-            "results['mis_ranked'] missing/empty. Required to color MIS."
-        )
-
-    best_rank = min(m.get("rank", 10**9) for m in mis_ranked)
-    best_mis_entry = next(
-        m for m in mis_ranked if m.get("rank", 10**9) == best_rank
-    )
-    mis1_ids = _extract_mis_nodes_1based(best_mis_entry, M) # These are 1-based indices
-    mis1 = [nodes[i-1] for i in mis1_ids] # Map to actual node names (could be strings)
-
-    if len(mis1) == 0:
-        keys = sorted(best_mis_entry.keys()) if isinstance(best_mis_entry, dict) else []
-        raise ValueError(
-            "Rank1 MIS came empty after canonical extraction. "
-            "This means the pipeline is generating empty MIS (or with values outside 0..M-1 / 1..M). "
-            f"rank1={best_rank}; rank1 item keys: {keys}"
-        )
-
-    mis1_set = set(mis1)
-
-    # --- graph (nodes 1..M) ---
-    G = nx.Graph()
-    G.add_nodes_from(nodes)
-
-    preserved_edges = []
-    removed_edges = []
-    for i in range(M):
-        for j in range(i + 1, M):
-            u_name = nodes[i]
-            v_name = nodes[j]
-            if A[i, j] != 0:
-                preserved_edges.append((u_name, v_name))
-                G.add_edge(u_name, v_name)
-            else:
-                removed_edges.append((u_name, v_name))
-
-    density = nx.density(G)
-
-    # layout + anti-overlap
-    pos = nx.spring_layout(
-        G,
-        seed=7,
-        k=3.0 / np.sqrt(max(M, 1)),  # Slightly larger k for more separation
-        iterations=1000,             # More iterations for better convergence
-        scale=1.0                    # Explicit scale to fill the plot area
-    )
-    pos = _enforce_min_distance(pos, min_dist=min_dist, iters=1200, seed=7)
-
-    fig, ax = plt.subplots(figsize=figsize)
-
-    # removed (subsample)
-    if show_removed and removed_edges:
-        draw_removed = removed_edges
-        max_removed_edges = 350 # Hardcoded for now, was a parameter
-        if max_removed_edges is not None and len(draw_removed) > max_removed_edges:
-            step = max(1, len(draw_removed) // max_removed_edges)
-            draw_removed = draw_removed[::step][:max_removed_edges]
-
-        nx.draw_networkx_edges(
-            G, pos,
-            edgelist=draw_removed,
-            style="dashed",
-            edge_color="0.65",
-            width=0.9, # removed_width was a parameter
-            alpha=0.45,
-            ax=ax,
-        )
-
-    # neighbors of Rank1 MIS
-    neigh_set = set()
-    for u_mis_idx in mis1_ids:
-        # A is 0-based
-        neighbor_indices = np.where(A[u_mis_idx - 1] != 0)[0]
-        for k in neighbor_indices:
-            neigh_set.add(nodes[k])
-    
-    mis1_set = set(mis1)
-    neigh_set -= mis1_set
-
-    # Separate edges for coloring
-    green_edges = []
-    other_preserved_edges = []
-
-    for u, v in preserved_edges:
-        if u in mis1_set or v in mis1_set:
-            green_edges.append((u, v))
-        else:
-            other_preserved_edges.append((u, v))
-
-    # Draw other preserved edges
-    if other_preserved_edges:
-        nx.draw_networkx_edges(
-            G, pos,
-            edgelist=other_preserved_edges,
-            edge_color="0.10",
-            width=1.15, # edge_width was a parameter
-            alpha=0.85,
-            ax=ax,
-        )
-
-    # Draw green edges
-    if green_edges:
-        nx.draw_networkx_edges(
-            G, pos,
-            edgelist=green_edges,
-            edge_color="C2",  # Green
-            width=1.15, # edge_width was a parameter
-            alpha=0.95,
-            ax=ax,
-        )
-
-    # nodes
-    node_colors = []
-    node_border_colors = []
-    label_colors = []
-
-    for u in nodes:
-        if u in mis1_set:
-            node_colors.append("C2")  # Green for Rank 1 MIS
-            node_border_colors.append("k")
-            label_colors.append("white")
-        elif u in neigh_set:
-            node_colors.append("k")  # Black for neighbors of Rank1 MIS
-            node_border_colors.append("k")
-            label_colors.append("white")
-        else:
-            node_colors.append("white") # Fallback for disconnected nodes if any
-            node_border_colors.append("k")
-            label_colors.append("black")
 
 
-    nx.draw_networkx_nodes(
-        G, pos,
-        nodelist=nodes,
-        node_size=420, # node_size was a parameter
-        node_color=node_colors,
-        edgecolors=node_border_colors,
-        linewidths=1.2,
-        ax=ax,
-    )
-
-    # Labels
-    for k, u in enumerate(nodes):
-        x, y = pos[u]
-        current_label_color = "white" if (u in mis1_set or u in neigh_set) else "black"
-        ax.text(
-            x, y, str(u), ha="center", va="center", fontsize=9, color=current_label_color, zorder=10 # font_size was a parameter
-        )
-
-    if title is None:
-        title = f"Graph — density={density:.2f} | Rank1 green | Neighbors black"
-    ax.set_title(title)
-    ax.axis("off")
-    plt.tight_layout()
-
-    return {
-        "mis_rank1_first": list(mis1),
-        "neighbors_of_mis": sorted(neigh_set),
-        "density": density,
-        "n_preserved": len(preserved_edges),
-        "n_removed": len(removed_edges),
-        "rank1": best_rank,
-        "fig": fig,
-        "ax": ax,
-    }
 
 
 # Stats / Alpha / Regime (reexported from misda._statistics)
@@ -419,91 +99,6 @@ def plot_custom_misda_graph(
 # MOP (Multi-Objective Pruning) - aka "Reduction" Helpers (for validation)
 # -------------------------------------------------------------------------
 
-def explain_ses(out, top_k=8, name=None, show_all=False):
-    """
-    Explains the result of calculate_ses(out). Returns string report.
-    SES = Structural Evidence Score.
-    """
-    if out is None or not isinstance(out, dict):
-        return "explain_ses: 'out' is invalid (expected dict)."
-
-    lines = []
-    def _p(x): lines.append(str(x))
-
-    title = f"Structural Evidence Score for {name}" if name else "Structural Evidence Score"
-    _p("\n" + " " * 72)
-    _p(title)
-    _p("-" * 72)
-
-    status = out.get("status", None)
-    mis = out.get("mis_size", None)
-    if mis is not None:
-        _p(f"Surrogate size (mis): {mis}")
-
-    if status == "NO_REDUCTION":
-        _p("Status: NO_REDUCTION (All objectives kept; reconstruction N/A).")
-        _p("SES = N/A  |  F_real = N/A  |  F_null = N/A")
-        return "\n".join(lines)
-
-    ses = out.get("ses", None)
-    F_real = out.get("F_real", None)
-    F_null = out.get("F_null", None)
-    r2_by_target = out.get("r2_real", None)
-
-    if ses is None or F_real is None or F_null is None:
-        _p("Status: UNDEFINED or missing metrics ('ses', 'F_real', 'F_null').")
-        _p("SES = N/A  |  F_real = N/A  |  F_null = N/A")
-        return "\n".join(lines)
-
-    gap = F_real - F_null
-    denom = max(1e-15, (1.0 - F_null))
-    ses_recalc = np.clip(gap / denom, 0.0, 1.0)
-
-    _p(f"SES = {ses:.4f}  (recalc = {ses_recalc:.4f})")
-    _p(f"F_real = {F_real:.4f}  |  F_null = {F_null:.4f}  |  gap = {gap:.4f}")
-    _p("Operational interpretation (Structural Evidence Score):")
-    _p("  - SES≈1: surrogate reconstructs others very well, far above null.")
-    _p("  - SES≈0: surrogate does not reconstruct better than null; suspicious reduction.")
-    _p("  - intermediate values: some reconstruction, but there is relevant loss.")
-
-    if ses >= 0.9:
-        _p("Short read: strong SES (reduction tends to be safe for reconstruction).")
-    elif ses >= 0.7:
-        _p("Short read: moderate SES (reduction may work, but deserves checking).")
-    else:
-        _p("Short read: low SES (high risk of surrogate being too small).")
-
-    if isinstance(r2_by_target, dict) and len(r2_by_target) > 0:
-        items = list(r2_by_target.items())
-        items_sorted = sorted(items, key=lambda kv: (-(np.inf) if kv[1] is None else kv[1]))
-        items_sorted = [(k, (-np.inf if v is None else float(v))) for k, v in items_sorted]
-        items_sorted = sorted(items_sorted, key=lambda kv: kv[1])
-
-        worst = items_sorted[:min(top_k, len(items_sorted))]
-        best = items_sorted[-min(top_k, len(items_sorted)):] if len(items_sorted) > 1 else []
-
-        def _fmt_r2(v):
-            if v is None or np.isneginf(v):
-                return "N/A"
-            return f"{v:.4f}"
-
-        _p("\nWorst targets (lowest R² in test):")
-        for k, v in worst:
-            _p(f"  {k}: R² = {_fmt_r2(v)}")
-
-        if best:
-            _p("\nBest targets (highest R² in test):")
-            for k, v in reversed(best):
-                _p(f"  {k}: R² = {_fmt_r2(v)}")
-
-        if show_all:
-            _p("\nR² by target (all):")
-            for k, v in items_sorted:
-                _p(f"  {k}: R² = {_fmt_r2(v)}")
-    else:
-        _p("\nR² by target is not available.")
-
-    return "\n".join(lines)
 
 
 # --------------------------------------------------------------------------------------
@@ -1027,126 +622,3 @@ def _analyze_adaptive(
         bootstrap_config={"b_bootstrap": b_bootstrap, "seed": seed},
         adaptive_config={"name": name, "caution": caution, "ensure_coverage": ensure_coverage}
     )
-
-
-
-def compile_benchmark_summary(results_dict, sort_by=None):
-    """
-    Standardizes the summary table generation for MISDA benchmarks.
-    Extracts all key metrics including Alpha regimes, Homogeneity, Linear/Non-Linear Fidelity, and Pareto Consistency.
-
-    Args:
-        results_dict (dict): Dictionary mapping Case Name -> MISDAResult object.
-                             Alternatively, can be a dictionary where values are dicts containing 'result_obj'.
-        sort_by (str): Column name to sort by.
-
-    Returns:
-        pd.DataFrame: Comprehensive summary table.
-    """
-    rows = []
-    
-    for case_name, item in results_dict.items():
-        # Handle both direct MISDAResult and wrapper dicts (e.g. from benchmark.ipynb)
-        # item could be MISDAResult or dict
-        res = None
-        truth_dim = None
-
-        if hasattr(item, 'best_mis'):
-            res = item
-        elif isinstance(item, dict) and 'result_obj' in item:
-            res = item['result_obj']
-            truth_dim = item.get('truth', {}).get('structural_expected' if _CORRELATION_MODE == "positive" else 'latent_expected', item.get('truth', {}).get('intrinsic_dim_expected', None))
-        
-        if res is None:
-            continue
-            
-        # Basic Params
-        # Handle res.Y being dataframe or numpy
-        N, M = res.Y.shape
-        algo_alpha = res.alpha
-        
-        # MIS Info
-        mis_indices = res.best_mis.indices if res.best_mis else []
-        dim_red = len(mis_indices)
-        
-        # Fidelity (Linear)
-        fidel_lin = None
-        if res.ses_results and isinstance(res.ses_results, dict):
-            fidel_lin = res.ses_results.get("F_real", None)
-             
-        # Fidelity (Non-Linear)
-        fidel_nl = None
-        try:
-            if N <= 5000: # Per safeguard
-                nl_out = calculate_ses_nonlinear(res.Y, mis_indices, n_estimators=50, return_details=True)
-                if isinstance(nl_out, dict):
-                    fidel_nl = nl_out.get("F_real", None)
-                else:
-                    fidel_nl = nl_out
-        except Exception:
-            pass
-             
-        # Pareto Consistency
-        prec, rec = 0.0, 0.0
-        try:
-             prec, rec = evaluate_pareto_consistency(res, res.Y)
-        except Exception:
-             pass
-             
-        # Homogeneity
-        homog = res.homogeneity_ratio
-        
-        # Status
-        status = "OK"
-        low_lin = (fidel_lin is not None and fidel_lin < 0.9)
-        low_nl = (fidel_nl is not None and fidel_nl < 0.9)
-        if low_lin and low_nl:
-            status = "LOW_FIDEL"
-        if prec < 1.0:
-            status = "UNSAFE(Prec)"
-        if truth_dim and dim_red != truth_dim:
-             status += f"|DimMismatch({dim_red}!={truth_dim})"
-        
-        # Alpha Bounds
-        a_min = res.alpha_min if hasattr(res, 'alpha_min') else 0.0
-        a_max = res.alpha_max if hasattr(res, 'alpha_max') else 1.0
-
-        def _fmt_f(val):
-            return f"{val:.2f}" if val is not None else "N/A"
-
-        row = {
-            "Case": case_name,
-            "Regime": res.regime.name if res.regime else "N/A",
-            "N": N,
-            "M": M,
-            "Dim(Red)": dim_red,
-            "Alpha": f"{algo_alpha:.4f}",
-            "Min": f"{a_min:.2f}",
-            "Max": f"{a_max:.2f}",
-            "Homog": f"{homog:.2f}",
-            "Fidel(Lin)": _fmt_f(fidel_lin),
-            "Fidel(NL)": _fmt_f(fidel_nl),
-            "Prec": f"{prec:.2f}",
-            "Rec": f"{rec:.2f}",
-            "Status": status
-        }
-        
-        if truth_dim is not None:
-            row["Exp"] = truth_dim
-            
-        rows.append(row)
-        
-    if not rows:
-        return pd.DataFrame()
-
-    df = pd.DataFrame(rows)
-    
-    # Reorder columns if Exp exists
-    cols = ["Case", "Regime", "N", "M", "Exp", "Dim(Red)", "Alpha", "Min", "Max", "Homog", "Fidel(Lin)", "Fidel(NL)", "Prec", "Rec", "Status"]
-    existing_cols = [c for c in cols if c in df.columns]
-    df = df[existing_cols]
-    
-    if sort_by and sort_by in df.columns:
-        df = df.sort_values(by=sort_by)
-        
-    return df
