+++ misda/_metadata.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: 2025 Monaco F. J. <monaco@usp.br>
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+"""Package metadata shared by the public facade and result objects."""
+
+__version__ = "0.4.2"
